backend/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.5
beautifulsoup4==4.9.3
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==3.1.7
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-jet-reboot==1.1.1
django-model-utils==4.1.1
django-nine==0.2.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
elasticsearch==7.13.1
elasticsearch-dsl==7.3.0
fake-useragent==0.1.11
feedparser==6.0.2
geographiclib==1.50
geopy==2.1.0
googletransx==2.4.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pycares==4.0.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-socks==1.2.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
schedule==1.1.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
yarl==1.6.3
